# feat(jsonpointer): use jx iterator API to return early