Benchstat result:
```
name       old time/op    new time/op    delta
Resolve-4    3.64µs ± 2%    3.27µs ± 2%  -10.03%  (p=0.000 n=14+15)
```